requirements.txt
# Trading & Exchange
ccxt==4.2.25
python-binance==1.0.19

# AI & Machine Learning
openai>=1.0.0
httpx>=0.24.0
ta==0.11.0
numpy==1.26.4
pandas==2.2.1

# Web Framework
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0

# Database
sqlalchemy==2.0.27

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Utilities
requests==2.31.0
python-dateutil==2.9.0
colorama==0.4.6
